Convert independent color verification script to pytest tests
